Update and rename TodoListApp to TodoListApp.py
 CODSOFT Python Programming Internship. 
Task-1: A To-do List Application in python using Tkinter GUI.
It is a basic practice of Tkinter GUI in python for developing a simple To-do list app. Basic level project. The to-dos are added to the list and marked green if they are completed. Then the ones finished can be deleted.